+++ src/jmcomic/jm_async_client.py
@@ -640,7 +640,7 @@ class AsyncJmApiClient(AsyncJmcomicClient):
                                with_ad_wcm=1,
                                need_total=True,
                                ) -> JmAlbumCommentPage:
-        """获取本子评论分页，返回 ``list`` 和 ``total``。"""
+        """获取本子评论分页，返回评论分页对象。"""
         resp = await self.req_api(
             self.API_FORUM,
             params={
@@ -649,7 +649,21 @@ class AsyncJmApiClient(AsyncJmcomicClient):
                 'aid': JmcomicText.parse_to_jm_id(jm_id),
             },
         )
-        return JmPageTool.parse_api_to_album_comment_page(resp.model_data)
+        return JmPageTool.parse_api_to_album_comment_page(resp.model_data, page)
+
+    async def forum_pagination(self,
+                               page=1,
+                               with_ad_wcm=1,
+                               ) -> JmAlbumCommentPage:
+        """获取全站评论分页。"""
+        resp = await self.req_api(
+            self.API_FORUM,
+            params={
+                'mode': 'all',
+                'page': page,
+            },
+        )
+        return JmPageTool.parse_api_to_album_comment_page(resp.model_data, page)
 
     async def add_favorite_album(self, album_id, folder_id='0'):
         """